server/requirements.txt
flask
flask-cors
transformers
torch
safetensors
pillow
